Adding customHook named as useFetchData
